requirements.txt
python-dotenv
azure-search-documents=11.4.0
langchain
openai
ragas
pandas
matplotlib
nest-asyncio